issue #53 fix: teamplanning rename
