+++ src/components/Planning/Bio/Bio.tsx
@@ -1,0 +1,13 @@
+import Calendar from "@/components/Planning/Calendar/Calendar";
+
+import { sectionStyle } from "@/components/Planning/Bio/Bio.style";
+
+const Bio = () => {
+	return (
+		<section css={sectionStyle}>
+			<Calendar />
+		</section>
+	);
+};
+
+export default Bio;
