+++ src/components/Planning/index.tsx
@@ -1,0 +1,5 @@
+import Bio from "@/components/Planning/Bio/Bio";
+import Calendar from "@/components/Planning/Calendar/Calendar";
+import Main from "@/components/Planning/Main/Main";
+
+export { Bio, Calendar, Main };
