--- src/components/TeamPlanning/index.tsx
@@ -1,5 +1,0 @@
-import Bio from "@/components/TeamPlanning/Bio/Bio";
-import Calendar from "@/components/TeamPlanning/Calendar/Calendar";
-import Main from "@/components/TeamPlanning/Main/Main";
-
-export { Bio, Calendar, Main };
