+++ src/pages/PlanningPage.tsx
@@ -1,0 +1,14 @@
+import Header from "@/components/Header/Header";
+import { Bio, Main } from "@/components/Planning";
+
+const PlanningPage = () => {
+	return (
+		<>
+			<Header />
+			<Bio />
+			<Main />
+		</>
+	);
+};
+
+export default PlanningPage;
