--- src/pages/TeamPlanningPage.tsx
@@ -1,14 +1,0 @@
-import Header from "@/components/Header/Header";
-import { Bio, Main } from "@/components/TeamPlanning";
-
-const TeamPlanningPage = () => {
-	return (
-		<>
-			<Header />
-			<Bio />
-			<Main />
-		</>
-	);
-};
-
-export default TeamPlanningPage;
