+++ src/pages/index.tsx
@@ -1,8 +1,8 @@
 import LandingPage from "./LandingPage";
+import PlanningPage from "./PlanningPage";
 import QuestionPage from "./QuestionPage";
 import SignUpPage from "./SignUpPage";
 import SirenDetailPage from "./SirenDetailPage";
 import SirenPage from "./SirenPage";
-import TeamPlanningPage from "./TeamPlanningPage";
 
-export { LandingPage, QuestionPage, SirenDetailPage, SirenPage, SignUpPage, TeamPlanningPage };
+export { LandingPage, QuestionPage, SirenDetailPage, SirenPage, SignUpPage, PlanningPage };
